requirements.txt
numpy>=1.19.0
pandas>=1.2.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tensorflow>=2.8.0
keras>=2.8.0
opencv-python>=4.5.0
Pillow>=8.0.0
jupyter>=1.0.0
notebook>=6.4.0
